// src/api/stetsom/model/auditEntry.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */
import type { AuditEntryAction } from "./auditEntryAction";

export interface AuditEntry {
  id: string;
  user_id: string;
  user_name: string;
  user_avatar?: string;
  action: AuditEntryAction;
  action_sentence: string;
  entity: string;
  entity_id?: string;
  entity_label?: string;
  created_at: string;
}