// src/api/stetsom/model/auditEntryAction.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */

export type AuditEntryAction =
  (typeof AuditEntryAction)[keyof typeof AuditEntryAction];

export const AuditEntryAction = {
  CREATE: "CREATE",
  UPDATE: "UPDATE",
  DELETE: "DELETE",
  LOGIN: "LOGIN",
  LOGOUT: "LOGOUT",
  PUBLISH: "PUBLISH",
} as const;